+++ terra/src/main/scala/com/dafttech/terra/engine/renderer/TileRendererBlock.scala
@@ -13,6 +13,6 @@ object TileRendererBlock {
 class TileRendererBlock extends TileRenderer {
   def draw(pos: Vector2, world: World, screen: AbstractScreen, tile: Tile, pointOfView: Entity, rendererArguments: AnyRef*) {
     val screenVec: Vector2 = tile.getPosition.toScreenPos(pointOfView)
-    screen.batch.draw(tile.getImage, screenVec.x + offset.x * BLOCK_SIZE, screenVec.y + offset.y * BLOCK_SIZE, BLOCK_SIZE, BLOCK_SIZE)
+    screen.batch.draw(tile.getImage, (screenVec.x + offset.x * BLOCK_SIZE).toFloat, (screenVec.y + offset.y * BLOCK_SIZE).toFloat, BLOCK_SIZE, BLOCK_SIZE)
   }
 }
